PR#25- Adding some changes
