+++ app/src/Components/Screens/Company/Company.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, StyleSheet, Text} from 'react-native';
 import type {PropsWithChildren} from 'react';
-import type {CompanyNavigationRoutesType} from '../../types/NavigationRoutesType';
+import type {CompanyNavigationRoutesType} from '../../../types/NavigationRoutesType';
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 
 type CompanyScreenProps = NativeStackScreenProps<
