+++ app/src/Components/Screens/Home/Home.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import {View, StyleSheet, Text, Button} from 'react-native';
-import Context from '../../Context/context';
-import type {AppContext} from '../../types/AppContextType';
+import Context from '../../../Context/context';
+import type {AppContext} from '../../../types/AppContextType';
 import type {PropsWithChildren} from 'react';
 import type {
   HomeNavigationRoutesType,
   NavigationRoutesType,
-} from '../../types/NavigationRoutesType';
+} from '../../../types/NavigationRoutesType';
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 import type {BottomTabScreenProps} from '@react-navigation/bottom-tabs';
 import type {CompositeScreenProps} from '@react-navigation/native';
@@ -15,7 +15,7 @@ type HomeScreenProps = CompositeScreenProps<
   NativeStackScreenProps<HomeNavigationRoutesType, 'HomeScreen'>,
   BottomTabScreenProps<NavigationRoutesType>
 > &
-  PropsWithChildren;
+  PropsWithChildren<any>;
 
 const HomeScreen: React.FC<HomeScreenProps> = ({navigation}) => {
   const {userProfile} = React.useContext(Context) as AppContext;
@@ -29,6 +29,10 @@ const HomeScreen: React.FC<HomeScreenProps> = ({navigation}) => {
         onPress={() =>
           navigation.navigate('ProfileStack', {screen: 'ProfileScreen'})
         }
+      />
+      <Button
+        title="Logout"
+        onPress={() => navigation.navigate('WelcomeScreen')}
       />
     </View>
   );
