+++ app/src/Components/Screens/Location/LocationSettings.tsx
@@ -1,24 +1,27 @@
 import React from 'react';
-import {View, StyleSheet, Text} from 'react-native';
+import {View, StyleSheet, Text, Button} from 'react-native';
 import type {PropsWithChildren} from 'react';
 import type {
-  SearchNavigationRoutesType,
+  LocationNavigationRoutesType,
   NavigationRoutesType,
-} from '../../types/NavigationRoutesType';
+} from '../../../types/NavigationRoutesType';
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 import type {BottomTabScreenProps} from '@react-navigation/bottom-tabs';
 import type {CompositeScreenProps} from '@react-navigation/native';
 
-type SearchScreenProps = CompositeScreenProps<
-  NativeStackScreenProps<SearchNavigationRoutesType, 'SearchScreen'>,
+type LocationSettingsProps = CompositeScreenProps<
+  NativeStackScreenProps<LocationNavigationRoutesType, 'LocationSettings'>,
   BottomTabScreenProps<NavigationRoutesType>
 > &
   PropsWithChildren;
 
-const SearchScreen: React.FC<SearchScreenProps> = () => {
+const LocationSettings: React.FC<LocationSettingsProps> = ({navigation}) => {
   return (
     <View style={styles.center}>
-      <Text>This is the search screen</Text>
+      <Text>This is the location setttings</Text>
+      <Text>You will be able to turn off and on your location here.</Text>
+
+      <Button title="Go Back" onPress={() => navigation.goBack()} />
     </View>
   );
 };
@@ -32,4 +35,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default SearchScreen;
+export default LocationSettings;
