+++ app/src/Components/Screens/Locked/SignUp.tsx
@@ -1,31 +1,30 @@
 import React from 'react';
 import {View, StyleSheet, Text, Button} from 'react-native';
 import type {PropsWithChildren} from 'react';
-import type {WelcomeNavigationRoutesType} from '../../types/NavigationRoutesType';
+import type {WelcomeNavigationRoutesType} from '../../../types/NavigationRoutesType';
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 
-type LogInScreenProps = NativeStackScreenProps<
+type SignUpScreenProps = NativeStackScreenProps<
   WelcomeNavigationRoutesType,
-  'Login'
+  'Register'
 > &
   PropsWithChildren;
 
-const LogInScreen: React.FC<LogInScreenProps> = ({navigation}) => {
+const SignUpScreen: React.FC<SignUpScreenProps> = ({navigation}) => {
   return (
-    <View style={styles.logInScreenView}>
-      <Text style={styles.title}>Log In to RightPay</Text>
+    <View style={styles.signUpScreenView}>
+      <Text style={styles.title}>Sign Up for RightPay</Text>
+      <Text style={styles.text}>Email</Text>
       <Text style={styles.text}>Username</Text>
       <Text style={styles.text}>Password</Text>
-      <Button
-        title="Forgot Password"
-        onPress={() => navigation.navigate('ForgotPassword')}
-      />
+      <Text style={styles.text}>Repeat Password</Text>
+      <Button title="Log In" onPress={() => navigation.navigate('Login')} />
     </View>
   );
 };
 
 const styles = StyleSheet.create({
-  logInScreenView: {
+  signUpScreenView: {
     flex: 1,
     alignItems: 'center',
   },
@@ -40,4 +39,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default LogInScreen;
+export default SignUpScreen;
