+++ app/src/Components/SplashScreen.tsx
